Make both unboxed tuples and TH work (fixes #718)
This sets us back to having both HscInterpreed and LinkInMemory
enabled (turns out we depend on GHC seeing these when it generates
the Core, among other things). But when it comes time to call
loadModule, we tell it we're in HscNothing/NoLink mode, which
prevents it from erroring out on unboxed tuples. Unlike the previous
attempt at a fix, this also works in the (rare?) case that someone uses
both Template Haskell and unboxed tuples in the same project.