Tweaking circle.yml
1. GHC 7.10.3

If a fresh circleci build is started (no cache) the available ghc is
8.0.1, and the machine config does not seem to do anything.

Explicitly install ghc-7.10.3, and configure the build to use it.

2. tasty junit output

For some reason, the --xml file to capture the junit.xml output has to
exist before the tests run, otherwise it crashes with an error related
to canonicalizePath. Create this file before running the tests.