Fixes ebind-05.fq
For some ebind n, isConc [| v= n |] is True, so even if we eliminate a
defining constraint `p -> v = n` from the graph, it can still get added
to wConcs at the "end" of the worklist, and thrown at z3.

This solves that problem by trivially passing that constraint by adding
False to the environment.