Add re-exported vars to GhcSrc
This commit is similar in spirit to the one that added to giTcs the
re-exported `TyCon`s, in the sense that this commit does the same but
for re-exported identifiers (i.e. GHC's 'Var's), by adding those to the
'giDefVars'. This was necessary as 'makeGhcSpec' will "export" a given
signature in the final liftedSpec only if the relevant symbol is
exported and "shows up" in the giDefVars.

This was necessary for "wrapper modules" that simply re-exports
everything, like the ones in 'liquid-base' or similar.