Fake the UnitId in GHC.Interface
This commit tweaks GHC.Interface.makeDependencies to create a \"fake
UnitId\" when building a StableModule. Obviously from a point of view of
correctness this is totally wrong, but for the executable it doesn't
matter, as the executable makes no use out of the information given by
having a 'Module' in scope. All it matters is that having a fake (but
locally unique) 'UnitId ' enables the relevant 'StableModule' to be
unique, so that we can use it as a key for the 'TargetDependencies'.