Provisional fix for classBinds
This commit fixes a guard in 'classBinds' for GHC 8.10.1. The regression
was introduced by the fact that the 'Outputable' representation for
'~<[]>' changed between 8.6.5 and 8.10.1.

Having said that, this is a provisional fix as comparing the string
representation of a class is hairy and we should try to use the proper
GHC combinators for that.