Allow bare embeds and cmeasures into a LiftedSpec
This commit changes 'makeLiftedSpec0' so that we allow input 'embeds'
and 'cmeasures' to survive the lifting process. Something to bear in mind
in that probably we would have to adjust the format of the correspondent
types, as it might make no sense to talk about a 'LocSymbol' for
something serialised into an interface file.