Introduce a StableName
This commit eschew the use of a NameSet to keep the exported names for a
module. This is because unique-based data structures relies on the
'Unique' of the underlying data structure to be, well, unique, but this
is not guaranteed by GHC for things like rebuilds, for example. By using
this new StableName we ensure that everything that a "wrapper" module
exports gets correctly refined in the target LiftedSpec.