Add more granular options to Data.Text.Fusion.Internal
This commit adds two LIQUID options to Data.Text.Fusion.Internal,
namely:

--no-termination
--no-totality

This is due to the fact the old executable was checking a single target
file by "slurping" in all the related dependencies, so the LIQUID
options were applied "holistically". In this example, we had the
'Data.Text.Fusion.Common' test which was importing '.Internal', but the
LIQUID options were set on the former, not the latter, and now due to
the fact the source plugin compiles each module indpendently, we don't
have a way to "propagate" liquid pragmas.

To say things differently, now we behave with regards liquid options to
the same way GHC deals with language extensions pragmas; we can specify
some global ones by passing them as plugin options, or individually, on
a case-by-case, for each file.