Do not call saveResult for the plugin
This commit introduces a new function called 'checkTargetInfo' that does
virtually the same of 'liquidOne' but it doesn't call 'saveResult'. This
is because the latter is used by the old executable to cache the
file on disk to perfom compilation avoidance, something that the plugin
gets for free via the GHC plugin pipeline.