Fix .circleci/config.yml
stack is fairly picky when it comes to the set of flags and options used
for compilation. If we are not consistent, we might end up running
'stack test' with different flags than 'stack build --fast liquid-base'
which would cause the latter to /not/ be present inside the stack's
ghc-pkg-db database by the time we call the testsuite.

It also add a flag to disable new executable.