Minor tweaks to the logging verbosity
This commit changes the default verbosity of some logging lines,
effectively "demoting" some from "normal" to "loud", so that in the
former configuration LH would only print warnings and errors.

On top of this, we also adjust the new binary to print the banner only
in certain circumstances, to conform to the old output as much as
possible.