Add logging verbosity as a Config option
This commit adds a new field to the LH's `Config` data structure,
unsurprisingly-named `loggingVerbosity`, which allows LH to have a
better control over the granularity of logging.

This allows the source plugin to run in `Quiet` mode (which is now the
default) but the executable to be customised to run by default using the
`--normal` mode, to show things like the progress bar.