Support for ghcide
This commit moves most of the plugin logic down to the typechecking
phase, so that we can integrate LH with tools like ghcide, that do not
evaluate things any further than TcRn. Previously we would do almost
everything inside a Core Pass (for historical reasons) and that was not
enough for making ghcide work.