Fix parsing errors and tests
This commit fixes the liquidhaskell-parser tests by actually removing
the artifical "<test>:source:column" prefix from errors, to deal with
the fact 'ppError' has been modified to not pprint the `pos` associated
with the error. This is fine, because we would always print the position
when we call 'printWarning' and 'printError' via the new diagnostics
API.

On top of this, we also fix an abrupt `throw` we had in the Plugin in
case of a parsing error, to be compliant with the standard error/warning
reporting mechanism we have introduced.