Amend emapReft to resolve BTyCon(s)
This commit modifies the `emapReft` function by adding an extra `c -> c`
function which can be used to resolve things like `BTyCon` when
this function is called on a `BareType`.

Previously, we were resolving all the types as the "leafs" of the
`RType` recursive data structure while neglecting the type constructor
of an `RApp`. Consider:

```
  | RApp  {
      rt_tycon  :: !c
    , rt_args   :: ![RType  c tv r]
    , rt_pargs  :: ![RTProp c tv r]
    , rt_reft   :: !r
    }
```

When `c` was instantiated to be `BTyCon`, we were neglecting the
`Symbol` contained inside it (and recursively to each member of
`rt_args`) and this was causing some name-resolution failures like the
ones in T1688.