Try resolving AConLike for Symbols
In case we need to qualify a Symbol, we would call the `ResolveSym`
instance defined for a GHC.Var, which would look only at Identifiers
(i.e. only `Id`s in a TyThing).

This PR adds an extra fallback case to look also for identifiers
associated to AConLike.

In a nutshell, this allows certain type synonyms to be correctly
expanded _in the lib module_ rather than in the client module when it's
too late.