Add stricter check for constructors of datatype refinements.
Constructors of refinements have to match the Haskell datatype
constructors exactly, with the exception of empty datatype decls
being allowed for attaching a termination measure.

This prevents accidentally parsing a GADT constructor refinement
as an ordinary top-level declaration due to layout, because an
empty GADT would trigger the error message.

This also replaces the previously implemented weaker check that
constructors being refined must be a subset of the constructors for
the original Haskell datatype.